Big changes: a new SX127xRadio driver and HMLayer gets a state machine
I stopped using mbed's SX1276 library because I wasn't getting DIO
interrupts and wrote my own library.  Turns out the Dragino HW was
at fault.

Designed HMLayer's state machine to be a flat 5-state design.
The state machine commands the device to normally listen (RX Cont)
and always transition through a Settings state to apply outstanding
settings between each TX & RX.

+++ HeyMacCmd.cpp
@@ -45,7 +45,7 @@ bool HeyMacCmd::cmd_txt(char const * const txt, uint8_t const sz)
 {
     bool success = false;
     uint8_t *payld = nullptr;
-    uint8_t offset = _frm->get_frm_sz();
+    uint16_t offset = _frm->get_frm_sz();
 
     if (offset + 1 + sz <= HM_FRAME_SZ)
     {
@@ -63,7 +63,7 @@ bool HeyMacCmd::cmd_cbcn(uint16_t const caps, uint16_t const status)
 {
     bool success = false;
     uint8_t *payld = nullptr;
-    uint8_t offset = _frm->get_frm_sz();
+    uint16_t offset = _frm->get_frm_sz();
 
     if (offset + 1 + 4 <= HM_FRAME_SZ)
     {
